Use ChangeTurf if needed for turf del in adminverb

+++ code/modules/admin/verbs/SDQL_2/SDQL_2.dm
@@ -93,7 +93,12 @@
 
 				if("delete")
 					for(var/datum/d in objs)
-						qdel(d)
+						if (isturf(d))
+							// turfs are special snowflakes that explode if qdeleted
+							var/turf/T = d
+							T.ChangeTurf(world.turf)
+						else
+							qdel(d)
 						CHECK_TICK
 
 				if("select")
